% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuck.R
\name{fuck}
\alias{fuck}
\alias{shit}
\alias{oh}
\title{Search last error in Google}
\usage{
fuck()

shit()

oh()
}
\value{
invisible character - text of the last error
}
\description{
Just type 'fuck' or 'shit' in the console just after the error and we will google for you...
}
\examples{
\dontrun{
stop("my strange error")
fuck
}
}
